refactor(test): move all test tags helpers to e2e/utils.ts
